requirements/analyze_app.txt
plotly==5.6.0
psycopg2-binary==2.9.1
streamlit==1.5.1
wordcloud==1.8.1
